shop/requirements.txt
backcall==0.1.0
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2018.11.29
chardet==3.0.4
decorator==4.3.2
Django==1.11
django-ajax-selects==1.7.1
django-debug-toolbar==1.11
django-form-utils==1.0.3
django-js-asset==1.1.0
django-mptt==0.9.1
django-user-agents==0.3.2
django-wysiwyg-redactor==0.5.1
idna==2.8
ipython
ipython-genutils==0.2.0
jedi==0.13.2
liqpay==1.0.1
mysqlclient==1.4.6
parso==0.3.2
pexpect==4.6.0
phonenumbers==8.10.4
pickleshare==0.7.5
Pillow
prompt-toolkit==2.0.8
ptyprocess==0.6.0
Pygments==2.3.1
PyMySQL==0.9.3
pytz==2018.9
PyYAML==3.13
reportlab==3.5.13
requests==2.21.0
six==1.12.0
sorl-thumbnail==12.5.0
soupsieve==1.7.3
#sqlite3==0.0.0
sqlparse==0.2.4
traitlets==4.3.2
transliterate==1.10.2
ua-parser==0.8.0
urllib3==1.24.1
user-agents==1.1.0
wcwidth==0.1.7
XlsxWriter==1.1.8
